Replace unsafe narrowing casts with overflow-checked alternatives
Use boost::numeric_cast<> to guard against silent integer overflow in
narrowing conversions (long→int, size_t→int, uintptr_t→long). Fix an
incorrect stream-position failure check in annotate.cc that used
static_cast<int>(pos) < 0 instead of comparing against the sentinel
pos_type value. Widen the stream-offset arithmetic in expr.cc to avoid
truncation on large files. Reverse a comparison direction in filters.cc
and add a positivity guard in report.cc to avoid signed→unsigned
conversion of negative values.

Resolves #815.

+++ src/filters.cc
@@ -189,7 +189,7 @@ void truncate_xacts::operator()(post_t& post) {
     last_xact = post.xact;
   }
 
-  if (tail_count == 0 && head_count > 0 && static_cast<int>(xacts_seen) >= head_count) {
+  if (tail_count == 0 && head_count > 0 && xacts_seen >= static_cast<std::size_t>(head_count)) {
     flush();
     completed = true;
     return;
